+++ src/app/app.component.ts
@@ -1,12 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
-import { AVALIABLE_LANGUAJES, DATABASE_NAME } from './constants/constants';
+import { AVALIABLE_LANGUAJES, DATABASE_NAME, FILES } from './constants/constants';
 import { MAGIC_NUMBERS } from './constants/number.constants';
 import { CharacterController } from './controller/characterController';
 import { Character } from './model/character';
 import { getMockCharacter } from './services/character.service.mock';
 import { CharactersService } from './services/characters.service';
+import { JsonService } from './services/json.service';
 import { StorageService } from './services/storage.service';
 @Component({
   selector: 'app-root',
@@ -27,21 +28,12 @@ export class AppComponent implements OnInit{
     private translate: TranslateService,
     private storageService: StorageService,
     private characterService: CharactersService,
-    private router: Router) {}
+    private router: Router,
+    private jsonService: JsonService
+  ) {}
 
   async ngOnInit() {
-    const devideLanguaje = window.navigator.language.substring(0,2).toLowerCase();
-    if( AVALIABLE_LANGUAJES.includes(devideLanguaje) ){
-      this.translate.setDefaultLang( devideLanguaje );
-    }else{
-      this.translate.setDefaultLang('es');
-    }
-    const protoCharacters = JSON.parse( await this.storageService.get(DATABASE_NAME) );
-    this.characters = CharacterController.converToCharacters(protoCharacters);
-    this.characterService.setCharacters(this.characters);
-    if(this.characters.length === MAGIC_NUMBERS.N_0){
-      this.characters.push(getMockCharacter());
-    }
+    this.fetch();
   }
 
   public changelanguage(language: string): void{
@@ -55,5 +47,36 @@ export class AppComponent implements OnInit{
     if(this.router.url.substring(0,7) !== '/detail') {
       this.router.navigate(['/detail']);
     }
+  }
+
+  private async fetch() {
+    this.setTranslations();
+    this.loadCharacters();
+    this.loadGiftData();
+  }
+
+  private setTranslations() {
+    const devideLanguaje = window.navigator.language.substring(0,2).toLowerCase();
+    if( AVALIABLE_LANGUAJES.includes(devideLanguaje) ){
+      this.translate.setDefaultLang( devideLanguaje );
+    }else{
+      this.translate.setDefaultLang('es');
+    }
+  }
+
+  private async loadCharacters() {
+    const protoCharacters = JSON.parse( await this.storageService.get(DATABASE_NAME) );
+    this.characters = CharacterController.converToCharacters(protoCharacters);
+    this.characterService.setCharacters(this.characters);
+    if(this.characters.length === MAGIC_NUMBERS.N_0){
+      this.characters.push(getMockCharacter());
+    }
+  }
+
+  private loadGiftData(){
+    this.jsonService.getJsonData(FILES.GIFTS)
+    .subscribe(data => {
+      this.characterService.setGiftData(data);
+    });
   }
 }