+++ src/app/constants/icon.constants.ts
@@ -1,0 +1,4 @@
+export enum IconTypes {
+  UP = 'chevron-up-outline',
+  DOWN = 'chevron-down-outline'
+}