+++ src/app/controller/character.randomize.utils.ts
@@ -1,0 +1,4 @@
+export const getRandomGender = (): string => {
+  const options = ['M', 'F', 'U'];
+  return options[Math.floor(Math.random() * options.length)];
+};