+++ src/app/services/json.service.ts
@@ -1,0 +1,23 @@
+import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+import { map } from 'rxjs/operators';
+import { GiftData } from '../model/giftData';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class JsonService {
+
+  constructor(private http: HttpClient) { }
+
+  public getJsonData(jsonFile: string) {
+    return this.http.get(jsonFile).pipe(
+      map(data => {
+        return data as GiftData;
+      })
+    );
+  }
+
+
+}
+